Updated application files
Updated files allow for client (app) to connect to simple socket server. 

In regards to socket server, the server must always be listening as application opens and closes connection after each command is sent. In the future, there may be a better method for handling this.

In regards to the application, the IP address referenced must be changed for actual implementation (and testing) as this is the IP I used during testing. In actual device implementation, each motor controller will have their own IP address that will be hard coded into the application.